+++ app/src/main/java/com/example/skibslogapp/model/Position/PositionCallback.java
@@ -1,4 +1,4 @@
-package com.example.skibslogapp.model.Koordinat;
+package com.example.skibslogapp.model.Position;
 
 import android.location.Location;
